requirements.txt
flask
PyPDF2
langchain
chromadb
sentence-transformers
langchain-community
google-generativeai